Relocated common code between user adding and user editing tests.
