+++ server/resources_helper_test.go
@@ -1,5 +1,13 @@
 package server
 
+import (
+	"context"
+	"net/http"
+
+	"github.com/pborman/uuid"
+	"github.com/rjkroege/mocking"
+)
+
 const embeddedResourceForPasswdchg = `
 	IsAuthed: {{.DecodedCookie.IsAuthed}}
 	DisplayName: {{.DecodedCookie.DisplayName}}
@@ -16,6 +24,14 @@ const embeddedResourceForListusers = `
 	Querysuccess: {{.Results.Querysuccess}}
 	Diagnosticmessage: {{.Results.Diagnosticmessage}}
 `
+
+type testPattern struct {
+	urlargs string
+	statuscode int
+	tapeResponse interface{}
+	tapeRecord []interface{}
+	outputString string	
+}
 
 // resourceHelper installs the above set of constant resources in place
 // of the resources read from the site directory (or compiled in.)
@@ -36,3 +52,29 @@ func resourceHelper() func() {
 
 	return func() { Resources = stashedResources }
 }
+
+// addCookie augments req with the context data that specifies that the
+// user is allowed to view users from the admin dialog.
+func addCookie(req *http.Request, capability CapabilityType) *http.Request {
+	uuid := uuid.NewRandom()
+	// User does have the capability to view users.
+	usercookie := &UserCookie{
+		Uuid:        uuid,
+		Capability:  capability,
+		Displayname: "Homer Simpson",
+	}
+	return req.WithContext(context.WithValue(req.Context(),
+		UserCookieStateName, usercookie))
+}
+
+// makeUnderTestHandlerListUsers creates a listUsers structure that
+// uses a mock (tape based) implementation of PasswordIndex.
+func makeUnderTestHandlerListUsers(tape *mocking.Tape) *listUsers {
+	undertesthandler := &listUsers{
+		// Always use the embedded resource.
+		embr:         makeEmbeddableResource(""),
+		passwordfile: (*mockPasswordIndex)(tape),
+	}
+	return undertesthandler
+}
+
